% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biodiversityPL}
\alias{biodiversityPL}
\title{Polish Biodiversity Data}
\format{
\subsection{\code{biodiversityPL}}{

A data frame with 48,461 rows and 24 columns:
\describe{
\item{id}{Identification}
\item{catalogNumber}{Catalog Number}
\item{scientificName}{Scientific Name}
\item{vernacularName}{Common Name}
...
}
}
}
\source{
\url{https://www.gbif.org/occurrence/search?dataset_key=8a863029-f435-446a-821e-275f4f641165}
}
\usage{
biodiversityPL
}
\description{
A subset of data from Global Biodiversity Information Facility.
}
\keyword{datasets}
