% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maptime_module.R
\name{maptimeUI}
\alias{maptimeUI}
\title{MAP and TIME Module UI and SERVER functions}
\usage{
maptimeUI(id)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
MAP and TIME Module UI and SERVER functions
}
